chore: Update package.json scripts and dependencies
This commit updates the package.json file with changes to the scripts and dependencies sections. It modifies the build script to include a prebuild step for the @kduprey/db package, and updates the dev script to use the turbo@1 dev command. Additionally, it adds new devDependencies for barrelsby, concurrently, dotenv-cli, and turbo.